style: 删除无用api

+++ src/main/java/vip/gadfly/tiktok/open/common/IRetryableExecutor.java
@@ -16,7 +16,7 @@ public interface IRetryableExecutor {
      * 默认：1000ms
      * </pre>
      */
-    default long getRetrySleepMillis() {
+    default int getRetrySleepMillis() {
         return 1000;
     }
